[005_extend_symmetries], issue #5, Added BC treatement for R120
The 'Boundary' class on the 'geom_extractor' module has been updated in the method that determines the characteristics of the BC for the borders of the geometry being exported. The previous implementation has been changed from an if-else structure to a match-case structure based only on the lattice type of geometry. In addition, the treatement for the R120 symmetry case has been added; depending on the border position, it assigns the 'ROTATION' or 'TRANSLATION' BC to the border. Some lattice types of geometry still have a missing implementation.