[005_extend_symmetries], issue #5, Added 'TRANSLATION' BC treatment
The 'Boundary' class on the 'geom_extractor' module has been updated in the method that determines the characteristics of the BC for the borders of the geometry being exported. Now, the case of a cartesian lattice with 'TRANSLATION' type of BC on its borders is handled. In addition, for the lattice types of geometry already treated, the assignement of the 'BoundaryType' value has been added.