[005_extend_symmetries], issue #5, Added 'RA60' BC treatment
The 'Boundary' class now handles the 'RA60' case: the 'ROTATION' and 'TRANSLATION' BCs are assigned to the borders of a sixth of a lattice. In addition, a different treatement in the evaluation of the starting and delta angles of an arc-type edge has been included: now, the function that writes the characteristics of the edges in the TDT file uses the modulus operator to always have the angles expressed in the 0-360° range.