[006_transf_operations], issue #6, Corrected lattice rotation
The 'rotate' method of the 'Lattice' class has been corrected. Previously, some compound objects were not rotated; now, it is done, so that the correct rotated geometry can be shown. In addition, the lattice regions were not built properly as the corresponding regions for each cell were not rotated in accordance with the lattice. In this sense, the 'rotate_from_axis' method has been added both in the 'GenericCell' and the 'GenericSurface' classes to be able to rotate the cell/surface around an axis different from the one on the center.