[005_extend_symmetries], issue #5, Changed geometric layout for a sixth symmetry
The geometry layout that identifies a sixth of a lattice has been changed. Previously, the first sixth was taken (starting from the right side); now, the fifth sixth is taken (in anticlockwise direction). This choice was made because tests on the RA60 symmetry type showed that the tracking in DRAGON5 did not fail only when the fifth sixth was considered. In this sense, both the SA60 and RA60 types of geometry, which imply a 'SymmetryType.SIXTH' value, now consider the same part of the lattice.