[005_extend_symmetries], issue #5, Improved BC treatement
In the 'Boundary' class, the method that evaluates the border characteristics (i.e. BC type, axes and angle) has been changed in its implementation. Now, the match-case structure has been reworked to group those lattice types of geometry sharing the same BC type. The border axes are initialized to the coordinates of the border starting point, which is evaluated accordingly to its orientation. For cases of lattices with a 'TRANSLATION' BC type on all its borders, the border axes are re-evaluated accordingly to the specific type of geometry.