[009_cell_update_geom], issue #9, Modified the functionality for updating the cell geometry
The 'update_geometry' method of the 'GenericCell' class failed to successfully update the geometric layout of the cell. Now, this functionality has been revised and provided in terms of two methods, i.e. 'update_geometry' and 'update_geometry_from_face'. The first one is meant to be used in SALOME as it updates the cell with the currently selected shape. The second one has been added to support updating the cell geometry from a Python script by providing the face to update the cell with.