[011_update_cell_with_circles], issue #11, Corrected addition of circles
The methods of the 'GenericCell' to update the cell geometry layout failed in updating the cell face if the given shape contained circle regions. The problem was due to cut and common operations performed when updating the dictionaries of cell regions VS associated properties/sectorization options in the method that adds a circle region. Now, these dictionaries are re-built by assigning a default value to the added region. In the context of the methods that update the cell layout, the values for the region properties are assigned according to the 'Region' objects.