[011_update_cell_with_circles], issue #11, Set properties inheritance for added circles
In the 'GenericCell' class, the operation that handles the addition of a circle surface to the cell (method 'add_circle') used to assign a default value for the properties of each added circle region. Now the regions resulting from the overlapping of the cell with the added circle inherit the properties of the overlapped regions of the cell, if any properties have already been defined.