[011_update_cell_with_circles], issue #11, Revised cell updating feature
The method 'update_geometry_from_face' of the 'GenericCell' class has been revised so to use already present methods. In addition, the cell re-initialization is now handled with a specific private method; since the actions performed in this new method are the same when initializing the 'GenericCell' class instance, the '__init__' method has been updated as well.