[013_different_dim_cells], issue #13, Revised base class for geometric surfaces
The name of the base class for geometric surfaces has changed to 'Surface' and its methods have been revised. In particular, the 'build_face' method has been removed since not necessary: now, when instantiating any of the subclasses of 'Surface', the corresponding face is directly built. Changes to other modules reflect the new version of the 'geometry.py' module.