[013_different_dim_cells], issue #13, Added 'GenericSurface' and utility module
The 'GenericSurface' class has been added to the 'geometries.py' module to account for any 2D generic surface defined from a given face object. In addition, the 'utility.py' module has been added to provide functions to be applied all over the application. They are meant to support the construction and the analysis of the geometry layout. In this sense, the 'update_relative_pos' function has been moved there; changes in the other modules reflects this.