[013_different_dim_cells], issue #13, Added support for generic type of cells
In the 'cells.py' module, the superclass for all the different type of cells has changed name from 'GenericCell' to 'Cell'; the once present 'SalomeCell' changed name to 'GenericCell'. The latter is now a subclass of 'Cell', thus all the methods that need an implementation has been added. The 'GenericCell' class is meant to support the creation of a cell from any kind of shape in SALOME. Changes to the 'lattice.py' module reflect the new naming of the classes. In addition, a new method for getting the type of a GEOM shape has been added to the 'geom_interface.py' module.