[013_different_dim_cells], issue #13, Applied sectorization when updating the cell geometry
When updating the cell geometry from a given shape (method 'update_geometry_from_face'), now the cell face is initialized to the face of the given shape. After extracting the edges to apply to the cell and updating the dictionaries associating the properties and the sectorization options to the cell regions, the sectorization operation is applied. In addition, the function 'get_object_from_id' in the 'geom_interface' module now catches any exception derived from any missing SALOME object associated to the given entry ID; in this case, None is returned. This has been done to avoid raising an exception when determining if a shape has to be removed from the current SALOME study.