[013_different_dim_cells], issue #13, Added 'layer' concept to the lattice building
The lattice construction logic has been revised to address the need to assemble lattices by superimposing cells of different dimensions. When adding a new cell or ring of cells, it is assigned to a specific layer (either a new one or an existing one). Lattice construction from cells has been removed from the 'add_cell' method and is now performed when displaying the lattice in the SALOME viewer. Layers are traversed in reverse order to cut those below, and cells are removed if the superior layer overlaps their centers. New methods have been added to the 'Lattice' class to support the new logic, while others have been modified or removed.