[013_different_dim_cells], issue #13, Updated edges identification
Modifications have been introduced in the 'LatticeDataExtractor' class to support the geometric data collection from lattice with cells of different dimensions. When retrieving the lattice edges corresponding to the ones of each region, an exception could happen from the result of the SALOME 'GetInPlace' operation. Now, it is correctly handled. In addition, the methods of the 'LatticeDataExtractor' class have been ordered alphabetically. The 'overlapping_cells.py' test example has been included in the 'tests/examples' folder to test the new functionality of assembling lattices by superimposing cells of different dimensions.