[013_different_dim_cells], issue #13 Handled translation/rotation of cells in layers
When applying a translation or a rotation on the 'Lattice' instance, the operation was not performed for the cells in the lattice layers; now, this has been included. In addition, the rotation axis construction has been corrected: now, it is always perpendicular to the lattice face, no matter the lattice position in the XY 2D space.