[013_different_dim_cells], issue #13, Separeted layers assemblage from compound update
The method of the 'Lattice' class dealing with assembling all the layers, eventually cutting the overlapped cells, previously also dealt with updating the lattice compounds from the updated cells. Now, the latter functionality has been separeted from this method and a dedicated one has been implemented. The code has been updated so that the behaviour is still the same.