[015_add_types_to_analysis], issue #15, Added geometry and property types for analysis
The type of geometry for the lattice cells and of property associated to the lattice regions is now provided to the function 'analyse_and_generate_tdt' of the 'main.py' module. Both values are propagated down to the method 'build_faces' of the class 'LatticeDataExtractor', where the face analysis is performed and 'Face' objects are built accordingly.