[015_add_types_to_analysis], issue #15, Corrected face number-property value ID association
The 'Face' objects are built by passing the GEOM face object and a string indicating the value of the property type (previously it was declared as a list). The association between face number and property value unique ID has been corrected to reflect the change in the property value type (from list to string).