[013_different_dim_cells], issue #13, Changed logic for overlapping layers
Previously, when a superior layer was collapsed onto an inferior one, only the cells of the lower layer whose centres were overlapped by the upper layer were removed. However, this condition was not general enough, so the logic for removing cells has changed. Now, only cells that are completely overlapped are removed from the corresponding layer; those that are partially overlapped are simply cut and their geometry is updated. It should be up to the user to decide whether to keep or restore the cut cells; a dedicated method for restoring cells, thus removing any inner circle, must be implemented.