[013_different_dim_cells], issue #13, Added function for retrieving changed cells
The 'get_changed_cells' function has been added to allow users to retrieve the cells of a 'Lattice' instance that show a change in their geometry layout (in terms of the area) wrt to their corresponding 'Surface' object. This can be used to get cells that are cut due to overlapping a superior layer of cells.