[026_restore_cell], issue #26, Added cell's restore functionality
The 'restore' method has been added to the 'Cell' class. This restores the cell's geometry layout to a state without any circular region with a face that reflects the current state of the cell's geometry. It also re-initialize the dictionaries of the cell's properties and sectorization options. Private methods for initializing the cell's geometry and dictionaries have been introduced as well.