[026_restore_cell], issue #26, Added restore lattice cells
The 'restore_cells' method has been included in the 'Lattice' class. It allows to restore the geometry layout of the given cells (copies of the ones in the lattice) to a state without any circular region. It also sets the cells' properties according to the provided dictionary of property type VS value.