[016_extract_regions], issue #16, Added flag for updating regions
A boolean flag has been added to the 'Lattice' class to specify the need to update the lattice geometry layout. Every method that modifies the geometry set this flag to True, whereas, when the regions are built (for displaying them in the SALOME viewer or exporting the TDT file) it is set to False. In this way, regions are rebuilt only if needed.