[016_extract_regions], issue #16, Added rebuilding of regions during lattice analysis
A call to the 'build_regions' method has been added during the initialization of the 'LatticeDataExtractor' class. When regions are not updated yet or the cells' geometry type is different from the one used to display the lattice regions, this method is called to update the lattice compound and build the list of 'Region' objects needed during the analysis of the lattice. In addition, attributes of the 'LatticeDataExtractor' class no more needed have been removed.