[016_extract_regions], issue #16, Added geometry type to compound evaluation
Previously, when building the lattice's regions, only the ones coming from the technological geometries of the cells were generated, no matter the passed item of the GeometryType enumeration. In this sense, even if the sectorized geometry layout of the cells was to be considered, regions were built always from the technological one. Now, the issue has been fixed by passing the geometry type to the private method dealing with updating the lattice compound object. The same goes for the lattice box: since it is a 'Cell' object, it might also have a sectorization; now, the type of geometry is passed to the private methods extracting the box subfaces and assembling the box with the cells.