[018_left_right_faces], issue #18, Changed faces identification method
When identifying if a face is on the left or on the right of an edge, the previous methodology, based on determining the minimum distance between a point on the left of the edge and the face being analysed, has been substituted with a new one. This allows to determine if the point is within or on the face border by a specified tolerance.