[017_lower_left], issue #17, Added lattice translation during analysis
The initialization of the 'LatticeDataExtractor' class has been changed to enable the translation of the lattice compound object, the corresponding 'Region' objects and the lattice center. This is done so that the lower-left corner of the geometry layout is positioned in the origin of the XYZ space. This operation is performed only for specific types of symmetries, and only if not already done beforehands. A dedicated function in the 'utility.py' module has been added for this purpose: it allows to translate a shape, having a reference point, so that the relative distance from a new reference point is kept the same.