[020_boundary_rev], issue #20, Code refactoring for 'geom_extractor.py'
In the 'geom_extractor.py' module some changes have been introduced. Regarding the 'Boundary' class, we have that: unnecessary attributes have been removed; the type of BC is no more provided as parameter, but directly evaluated during the instance initialization; the check on the BC type to be AXIAL_SYMMETRY or TRANSLATION has been removed as no more needed; the 'find_edges_on_border' method has been revised and now the indices of the edges belonging to a boundary are directly extracted from their names instead of needing an additional dictionary (a dedicated function has been added in the 'utility.py' module). As for the 'LatticeDataExtractor' class, we have that: the 'Lattice' instance is now deep copied to avoid unwanted modifications; the 'build_boundaries' method has been changed to support the modifications made in the 'Boundary' class.