[021_circles_check], issue #21, Modified check on circles' presence
When adding a circle to the 'Cell' instance, a check on the same circle's presence was done by searching among the cell's centered ones only. Now, the check takes also the circle's position so to consider also any not centered circles. The bug on the result of this check, that was present in the 'add_circle' method, has been corrected. In addition, the 'remove_circle' method has been modified to enable the possibility of removing circles at any position within the cell. The corresponding private methods have been updated to comply with the changes.