[022_set_type_geo], issue #22, Added method for setting the lattice type of geometry
The method 'set_type_geom' has been included in the 'Lattice' class. It allows to set the lattice type of geometry to a value that is valid for the types of cells, the applied geometry and eventually the presence of a box. To support this functionality, a dictionary associating for each type of cells, the valid combinations of types of symmetry and lattice types of geometry has been included in the 'support.py' module.