[031_even_cells], issue #31, Modified addition of a ring of cells
The method 'add_ring_of_cells' of the 'Lattice' class has been modified to support cartesian lattices having an even number of cells on their side. The dimensions of the construction figure, where cells are added along its borders, are evaluated on the basis of a multiplication factor which depends on whether the lattice is built from a central cell or not.