[031_even_cells], issue #31, Added property for rotation angle of cells
In the 'Lattice' class, the attribute representing the rotation angle of the main pattern of cells has been transformed into a property with only a getter method. When adding a cell or a ring of cells, the rotation angle of the main pattern of cells is evaluated by considering the most common value among the cells in the lattice.