[031_even_cells], issue #31, Modified ring of cells construction
If the central cell has not yet been defined when adding a ring of hexagonal cells, the method failed to assemble the cells properly. Now, the correct evaluation of the multiplication factor that determines the dimensions of the construction figure along whose borders the cells are placed is now provided.