[031_even_cells], issue #31, Modified lattice box construction
The logic for building a lattice box has been updated by considering the bounding box from the compound of cells. Its dimensions are used to evaluate the characteristic dimensions of the figure the box corresponds to (either a rectangle or a hexagon). Now, boxes can also be correctly constructed for lattices with an even number of cartesian cells.