[024_test_suite], issue #24, Added modifications to surface classes
Modifications to support the instantiation of 'Surface'-based classes have been introduced: these include the implementation for enabling rectangular shapes with rounded corners and checks on the proper shape type used when updating the face object. To support the latter, a specific function has been added to the 'utility.py' module.