[024_test_suite], issue #24, Added unittest structure and first test classes
The structure for running GLOW's unittests has been defined. The 'test/unittest' folder contains a 'run_tests.py' module to be provided to the SALOME shell so that all the tests are correctly run within the SALOME framework. In the 'test_geometries.py' module, two classes have been provided: one is abstract, 'TestSurface', and its test methods are meant to be run only by its subclasses, one of which is provided. The subclass 'TestCircle' allows to test the implementation related to the 'Circle' class.