[024_test_suite], issue #24, Applied changes to support tests
Modifications have been introduced in the 'Cell' class to support testing the methods that are intended to be run from the SALOME GUI. Now, private methods have been extracted so that the implemented logic can be tested. In addition, a bug on the update of the cell's dictionaries have been corrected.