[024_test_suite], issue #24, Added test to check 'restore_cells' and 'set_type_geo' methods
The tests for the methods 'restore_cells' and 'set_type_geo' of the 'Lattice' class are now present. In addition, modifications have been included in the 'Lattice' class to support restoring all the indicated cells, not only those whose circular regions have been cut. The 'Cell' objects in the 'lattice_cells' attribute are now directly linked to the ones in the 'layers' attribute.