[024_test_suite], issue #24, Added tests for functions and 'GenericCell' class
Tests for the functions declared in the 'cells.py' and 'lattices.py' modules have been added, as well as those for the 'GenericCell' class. Missing implementations for the 'GenericClass' methods have been included. A correction in the method that updates the cell's geometry layout has been implemented: now, the edges representing the borders of the shape to update the cell with are excluded from the analysed ones. In the 'Lattice' class, the method 'sectorize' has been removed as its implementation is no more necessary: in fact, the sectorization can be applied to the cells themselves and the result displayed with the 'show' method specifying 'SECTORIZED' as the geometry type to show.