[024_test_suite], issue #24, Added tests for 'Boundary', 'Edge' and 'Face' classes
Tests for the 'Boundary', 'Edge' and 'Face' classes have been added in the 'test_geom_extractor.py' module. In addition, modifications have been introduced in these classes to support their testing. Corrections have been introduced in the method that determines the BC type in the 'Boundary' class. Now, lattices with type 'ROTATION' (typgeo=2), used with a TISO tracking, are considered, and the associated BC types are 'ROTATION' and 'TRANSLATION'. For a cartesian lattice with type 'RECTANGLE_TRAN' (typgeo=5), the border's axis directions were calculated wrongly; now, it has been corrected.