[040_transl_symm], issue #40, Handled automatic lattice repositioning
The automatic translation of the lattice to analyse could fail when the user had already translated the lattice into the correct position. The lower-left position check has been fixed by specifying a tolerance, so to avoid unwanted translations. Additionally, in all other lattice cases where the user translated the lattice, now the lattice elements are properly translated so that the center coincides with the XYZ origin.