[039_lattice_box], issue #39, Added direct box assignment
In the 'Lattice' class, the 'lattice_box' attribute has been transformed into a property. This change allows to directly set the box with a 'Cell' object without calling the 'build_lattice_box' method that requests to indicate the thicknesses of the box layers. In addition, if None is provided to the property, the box is removed. The problem for which a box could not be shown when setting the lattice box to a cell has been solved by introducing a check on the box presence in the method that assembles the box with the cells.