[024_test_suite], issue #24, Added tests for checking functions of the 'geom_extractor.py' module
Tests for the functions of the 'geom_extractor.py' module have been added together with one for the private method '__get_unique_edges' of the 'LatticeDataExtractor' class. In addition, some modifications have been provided to the docstrings in the modules 'geom_extractor.py' and 'geom_interface.py'. In the 'LatticeDataExtractor' class, when extracting the lattice's edges to analyse during the initialization, common edges are now only considered once.