[024_test_suite], issue #24, Added unittest for the 'generator.py' module
Unittest have been included for the 'TdtData' dataclass and the functions that write the output TDT file. Modifications to the docstrings and comments are present. In addition, the type of geometry of the lattice for a third and sixth symmetries and TISO tracking type has been corrected: the previous 'SYMMETRY_TWO' value has been changed to 'ROTATION' for a third symmetry, whereas the same has been added for the sixth symmetry, as it was missing.