[019_tdt_file], issue #19, Added the albedo setting
The function 'analyse_and_generate_tdt' of the 'main.py' module, requires now an object of the 'TdtSetup' dataclass; this stores the settings for configuring the geometry TDT export. In particular, it allows to set the value of the albedo for the BCs so that, for an ISOTROPIC typgeo, any value in the 0-1 range can be set, whereas for typgeo > 0 the only accepted value is 0.0. Proper modifications to the involved parts have been provided to support the new implementation.