[019_tdt_file], issue #19, Changed values in TDT file to scientific notation
Values indicating the geometric characteristics of the edges and BCs are now expressed in scientific notation in the TDT file. Two constants have been added to the 'generator.py' module, one indicating the number of digits after the decimal, the other the format to use for writing values in the TDT file.