[024_test_suite], issue #24, Modified tests to comply with latest changes
Test cases already defined have been modified to comply with the new implementation merged from the branch 'master'. Now, the tests of the initialization of the 'TdtData' dataclass include cases to assess the correct behaviour when different combinations of the albedo and the lattice's type of geometry are provided. Since the format with which geometric data is written in the output TDT file has changed, the corresponding tests have been updated. In addition, given the newly added 'TdtSetup' dataclass, tests of its initialization have been included as well.