[014_user_manual], issue #14, Updated documentation
The documentation has been updated by adding the user manual chapter and the bibliography one. In the folder 'docs/sources/images', images about the GLOW usage included in the user manual chapter are now present. The chapter introducing the tests have been included. The RST files of the API guide have been updated to reflect the current state of the code. The 'index.rst' and the 'conf.py' have been updated to comply with the new modifications.