[014_user_manual], issue #14, Applied modifications to the doc and README
The introductory section of the 'Getting Started' chapter has been moved to the 'Introduction' chapter substituting  what was already available. The dependencies have been updated both in the doc and in the README. The latter has been modified as well by reducing the content of its 'Introduction' section together with other small changes.