[047_toml], issue #47, Added toml
A toml file has been added to the project. Now, GLOW can be installed with pip and its modules are visible since they are imported directly when importing GLOW in a Python script. The documentation and the README have been updated to reflect the new state of the project; they explain how to install and import modules. The functional tests have been updated by removing the need to add the modules of GLOW to the PYTHONPATH. The .gitignore file has been updated to ignore the 'build' folder.