[051_multiple_lattices], issue #51, Added preliminary colorset treatment
The analysis functionality has been reworked to allow the generation of a TDT file from a block of lattices or from a colorset. The 'analyse_and_generate_tdt' function now requires a list of 'Lattice' objects to analyse, so that layouts made by multiple lattices can be exported to TDT. In addition, a compound object can be provided: if so, the analysis is perfomed on it, given the list of lattices which are needed to link the compound's faces to the regions' properties.