[051_multiple_lattices] Added typegeo and symmetry type at TDT export
The 'TdtSetup' dataclass has been modified to include the indication of the typegeo and symmetry type of the layout exported to TDT. Values provided to this data structure when calling the 'analyse_and_generate_tdt' function overwrites any values previously defined automatically by GLOW itself. This possibility has been added to support the TDT export of colorset, or their portion, as they are not associated to a class in GLOW. Similarly, when exporting a 'Lattice', users now need to provide values of typegeo and symmetry type that match the lattice's layout. The 'colorset.py' example has been added to the 'tutorials/' folder to show how to build and export a portion of a colorset.