[051_multiple_lattices], issue #51, Updated 'geom_entities.py' module and angle calculation at the analysis step
The docstrings of the wrapper classes in the 'geom_entities.py' module have been updated or included if missing. A specific implementation of the '+' operator has been included for both 'Compound' and 'Face' classes to take into account the need to build a GEOM face as a result of the fusion. In addition, a correction on the calculation of the angle of the boundary has been included in the 'Boundary' class: now, an absolute tolerance of 1e-6 has been specified, so that the value is set to zero if lesser than this tolerance.