[051_multiple_lattices], issue #51, Added module 'layouts.py'
The new module 'layouts.py' has been included in the repository. It holds the classes enabling the creation and visualisation of any geometry layouts. Currently, it includes the abstract class 'Layout', which provides attributes and methods common to all geometry layouts (i.e. cells and lattices). In addition, a new implementation of the 'Region' class is present. Contrary to the previously implemented 'Region' class, now we have that 'Region' is a GEOM face, meaning that it behaves like any GEOM face when used with the GEOM functions. Moreover, '+' and '-' overloads have been included to manage the fuse and cut operations among 'Region' objects. The goal of this new implementation is to use 'Region' instances not only for visualisation purposes, but also to enable a more intuitive way of defining the layout of a cell/lattice. The old 'Region' class is still kept for compatibility with existing code. It will be removed in future commits.