[051_multiple_lattices], issue #51, Added 'Layout' inheritance to 'Region'
The 'Region' class has been modified by adding inheritance from the 'Layout' abstract class. This was found needed to simplify the application of Euclidean transformations, as the corresponding logic is now included in the dedicated methods, whose signature is inherited from the 'Layout' class. In addition, the '*' operator has been overloaded to perform a common operation between two 'Region' objects; the same goes for the '==' operator of the 'Vertex' wrapper class to identify if two vertex are the same on the basis of the XYZ coordinates.