[051_multiple_lattices], issue #51, Added 'Fillable' class
The 'Fillable' class has been included in the 'layouts.py' module to take into account for any layout that can be filled by other geometry layouts. In the perspective of reproducing the hierarchical tree-like structure of a generic layout, the 'Fillable' class represents a node, while the 'Region' class a leaf. In this sense, a 'Fillable' object has a list of layers, where each layer can include a list of leaves or nodes. The order of the layers determines which layout objects will be cut when the entire layout is assembled; in this sense, cuts will be performed by traversing the layers in reverse order. Currently, only the capability to add a generic layout to the layers is present.