[051_multiple_lattices], issue #51, Added check in export function
The 'analyse_and_generate_tdt' function has been modified by adding a preliminary check on the correctness of the settings of the lattices or the data in the 'TdTSetup' instance. To realise it, a support function has been added in the 'utility.py' module. In addition, a further modification to the 'build_compound_borders' function has been included to take into account also for non-segment type of edges on the borders.