[051_multiple_lattices], issue #51, Updated analysis step
Some changes have been introduced in the class 'LatticeDataExtractor' mainly with the introduction of two attributes, one for the layout center and the other for the layout characteristic dimensions. When building the 'Boundary' instances, it is no more valid to consider the characteristic dimensions of the first lattice in the given list as the dimensions for the entire layout, since now the analysis can be performed on multiple lattices. In addition, this step should not rely on translating the center of the first lattice: the new attribute avoids this operation, while still referencing the center of the first lattice if there is no need to translate the lattice compound. Docstrings and comments have been updated appropriately, while the output of the 'get_bounding_box' function now is rounded to the sixth decimal digit to remove floating-point precision errors that SALOME could introduce.