[051_multiple_lattices], issue #51, Updated unittest for 'geom_extractor.py' module
The unit tests for the content of the 'geom_extractor.py' module have been updated to consider also the case in which the analysis prior to the TDT file generation comes from either a full colorset or a portion of it. The logic for building the 'Region' objects for a specified compound, that is part of the lattices, has been fixed to properly raise an exception when no region can be associated with any of the faces of the compound.