[051_multiple_lattices], issue #51, Modified 'Cell' class, utility function and 'main.py' unittest
The 'Cell' class has been modified to deal with the case in which a cell is updated with a shape having a hole. Previously, the cell was updated ignoring the hole with only the hole border added as a new edge to the cell's layout. Now, holes are properly considered due also to the modification to the partition operation performed in the 'add_circle' method. The 'build_compound_borders' function has been updated to ignore the collinear check for those edges of the compound borders which are not segments. Given these changes, the unit tests for the 'main.py' module, that test the generation of a TDT file from a colorset, have been updated by specifying the hash code of the two new reference files.