[051_multiple_lattices], issue #51, Updated functional tests
The functional tests have been updated to reflect the new approach to the generation of the TDT file, which requires to provide a list of lattices to the 'analyse_and_generate_tdt' function. The reference .dat files have been moved to a dedicated folder as well as the .dat files generated by running the functional tests. This has been done to provide a clearer structure of the tests. The 'run_test.py' script has been modified accordingly by adding more comments and slightly changing the logic: now, if a reference .dat is missing, the corresponding test is skipped before running its script. In addition, the log file is produced only if any test script fails in generating the .dat file. The 'run_test.py' script creates a folder dedicated to storing the output files of the tests and moves the .dat files in it after each test is performed.