[053_macro_regions], issue #53, Added capability to include macro regions
The need to include the macro regions data in the TDT file has been addressed to by including the 'MACRO' item in the 'PropertyType' enumeration. This new property can be assigned to the cell's regions in the same way as the 'MATERIAL' property type. When exporting the layout to the TDT file, users can now specify the list of 'PropertyType' items to include. This list defaults to the 'PropertyType.MATERIAL' value. The analysis step now is performed by taking into account all the indicated properties. If the 'PropertyType.MACRO' is included, the names of the macro geometries and the indices assigned to the regions of the layout are written in the TDT file; otherwise the default values are used.