[053_macro_regions], issue #53, Added update cell's properties method
The 'Cell' class has been updated by adding the new method 'update_properties' to enable users to update a subset of the property types assigned to the cell's regions without restoring the internal dictionary in the process. In addition, the 'get_region_info' function has been corrected to print the value for each of the property types assigned to the selected region of a cell/lattice.