[053_macro_regions], issue #53, Changed lattice rotation check
Previously, when a 'Lattice' instance was rotated, a check on the rotation angle of the cells of the lattice's main pattern was performed. This check assessed whether the cells' rotation angle was among the values of a list of admitted angles. However, for hexagonal cells, the indicated values were too restrictive: in fact, any multiple of 30° should be considered valid since this cell relies on a regular hexagon and any k*30° rotation (k>0, int) results in either a X- or an Y-oriented hexagon. To address this issue, the 'VALID_CELLS_ANGLES' constant has been removed and a variable added to the private method that handles the check. There, the variable is computed according to the cells' type: for hexagonal cells, the admitted values are the multiples of 30°, for Cartesian cells, they are the multiples of 90°.