[053_macro_regions], issue #53, Removed check on rotation
When rotating a 'Lattice' instance, a check on the rotation angle of its cells was performed. This check raises an exception whenever all the cells do not share the same rotation, which can limit the creation of layouts. Hence, it has been removed. In addition, in function 'build_compound_borders', the initial partition now sets the type of the resulting shape to be a face object as this solves problems in finding the borders of colorset layouts.