[052_modular_approach], issue #52, Added layers management
The 'Fillable' class has been modified to manage the layers it could be made of. The method 'build_regions' has been added to build a list of 'Region' objects representative of the technological geometry of the layout. These regions are the result of collapsing all the layers by traversing the entire hierarchical structure in reverse order, and recursively cutting the layers that are overlapped. In addition, helper functions has been added to the 'layouts.py' module to handle the assignment of colors to the regions according to the values of a given property type.