[052_modular_approach], issue #52, Added show() for 'Fillable'
The class 'Fillable' has been modified by adding the 'show()' method. It allows users to display objects of the 'Fillable' subclasses in the SALOME 3D viewer by adding their regions to the study. Regions are those of the technological geometry obtained by collapsing all the layers and associating a colour according to a mapping to an indicated property type. In addition, the abstract method 'show()' in the 'Layout' class and the implementation on its subclasses has changed by removing '*kwargs' among its arguments.