[052_modular_approach], issue #52, Implemented 'rotate' method in 'Fillable' class
The 'Fillable' class has been updated by adding the implementation for rotating any layout. In addition, the signature of the 'rotate' method in the abstract class 'Layout' has been corrected by adding the need for the axis parameter. Given this change, the corresponding method in the 'Region' class has been updated to enable the possibility to provide the rotation axis. An issue related to the case when cutting a 'Fillable' object (due to layers overlapping) resulted in only its regions to be updated has been solved. Now, the cut is applied to each layer and not directly to the regions, which are rebuilt afterwards.